requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0

# OpenAI
openai>=1.10.0

# Image processing (for pipeline)
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0
huggingface-hub>=0.20.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
python-dotenv>=1.0.0
